WIP: started going through all the post-worlds changes needed to be done to clean up the code; also fixed the camera translations
