+++ src/main/java/frc/robot/commands/Launch.java
@@ -4,18 +4,10 @@
 
 package frc.robot.commands;
 
-import static edu.wpi.first.units.Units.Inches;
-import static edu.wpi.first.units.Units.Meters;
-
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Translation2d;
-import edu.wpi.first.math.interpolation.InterpolatingTreeMap;
-import edu.wpi.first.math.interpolation.Interpolator;
-import edu.wpi.first.math.interpolation.InverseInterpolator;
-import edu.wpi.first.units.measure.Distance;
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.DriverStation.Alliance;
-import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants.ColumnConstants;
 import frc.robot.Constants.FieldConstants;
@@ -26,6 +18,9 @@ import frc.robot.subsystems.column.Column;
 import frc.robot.subsystems.hood.Hood;
 import frc.robot.subsystems.hopper.Hopper;
 import frc.robot.subsystems.shooter.Shooter;
+import frc.robot.util.ShotMapUtil;
+import frc.robot.util.ShotMapUtil.Shot;
+
 import java.util.function.Supplier;
 
 public class Launch extends Command {
@@ -36,54 +31,6 @@ public class Launch extends Command {
   private final Supplier<Pose2d> m_poseSupplier;
   private Shot m_shot;
   private final String m_position;
-  private final Timer m_timer = new Timer();
-  private double m_timeDelay;
-
-  // Lookup table mapping shooting distance in meters to Shot parameters (RPM + hood position).
-  // We use interpolation between known distances so we can smoothly compute
-  // appropriate shooter settings for any intermediate distance.
-
-  public static class Shot {
-    public final double m_shooterRPM;
-    public final double m_hoodPosition;
-
-    public Shot(double shooterRPM, double hoodPosition) {
-      m_shooterRPM = shooterRPM;
-      m_hoodPosition = hoodPosition;
-    }
-  }
-
-  private static final InterpolatingTreeMap<Distance, Shot> distanceToShotMap =
-      new InterpolatingTreeMap<>(
-          (startValue, endValue, q) ->
-              InverseInterpolator.forDouble()
-                  .inverseInterpolate(startValue.in(Meters), endValue.in(Meters), q.in(Meters)),
-          (startValue, endValue, t) ->
-              new Shot(
-                  Interpolator.forDouble()
-                      .interpolate(startValue.m_shooterRPM, endValue.m_shooterRPM, t),
-                  Interpolator.forDouble()
-                      .interpolate(startValue.m_hoodPosition, endValue.m_hoodPosition, t)));
-
-  static {
-    distanceToShotMap.put(
-        Inches.of(47.96), new Shot(ShooterConstants.HUB_SPEED_RPM, HoodConstants.HUB_SETPOINT));
-    distanceToShotMap.put(
-        Inches.of(122.12),
-        new Shot(ShooterConstants.TOWER_SPEED_RPM, HoodConstants.TOWER_SETPOINT));
-    distanceToShotMap.put(
-        Inches.of(134.13),
-        new Shot(ShooterConstants.TRENCH_SPEED_RPM, HoodConstants.TRENCH_SETPOINT));
-    distanceToShotMap.put(
-        Inches.of(192.0),
-        new Shot(ShooterConstants.CORNER_SPEED_RPM, HoodConstants.CORNER_SETPOINT));
-  }
-
-  public Distance getDistanceToHub() {
-    Translation2d robotPosition = m_poseSupplier.get().getTranslation();
-    return Meters.of(robotPosition.getDistance(m_target));
-  }
-
   private Translation2d m_target = FieldConstants.BLUE_HUB_POSITION;
 
   public Launch(
@@ -112,13 +59,8 @@ public class Launch extends Command {
             .map(a -> FieldConstants.RED_HUB_POSITION)
             .orElse(FieldConstants.BLUE_HUB_POSITION);
 
-    m_timer.restart();
-
     m_hopper.setVoltage(-HopperConstants.LAUNCHING_VOLTAGE);
-  }
 
-  @Override
-  public void execute() {
     if (m_position.equals("trench")) {
       m_shot = new Shot(ShooterConstants.TRENCH_SPEED_RPM, HoodConstants.TRENCH_SETPOINT);
     } else if (m_position.equals("hub")) {
@@ -127,25 +69,19 @@ public class Launch extends Command {
       m_shot = new Shot(ShooterConstants.TOWER_SPEED_RPM, HoodConstants.TOWER_SETPOINT);
     } else if (m_position.equals("corner")) {
       m_shot = new Shot(ShooterConstants.CORNER_SPEED_RPM, HoodConstants.CORNER_SETPOINT);
-    } else if (m_position.equals("relay")) {
-      m_shot = new Shot(ShooterConstants.RELAY_RPM, HoodConstants.K_MAX_POSITION);
     } else {
-      m_shot = distanceToShotMap.get(getDistanceToHub());
+      m_shot = ShotMapUtil.distanceToShotMap.get(ShotMapUtil.getDistanceToHub(m_poseSupplier, m_target));
     }
+  }
 
-    if (m_shot.m_hoodPosition < 0.3) {
-      m_timeDelay = 0.0;
-    } else {
-      m_timeDelay = 0.0; // 3.0
-    }
-
+  @Override
+  public void execute() {
     m_shooter.setVelocity(m_shot.m_shooterRPM);
     m_hood.setPosition(m_shot.m_hoodPosition);
 
     if ((m_shooter.getLVelocity() > (m_shot.m_shooterRPM - ShooterConstants.TOLERANCE_RPM))
         && (m_shooter.getMVelocity() > (m_shot.m_shooterRPM - ShooterConstants.TOLERANCE_RPM))
-        && (m_shooter.getRVelocity() > (m_shot.m_shooterRPM - ShooterConstants.TOLERANCE_RPM))
-        && (m_timer.hasElapsed(m_timeDelay))) {
+        && (m_shooter.getRVelocity() > (m_shot.m_shooterRPM - ShooterConstants.TOLERANCE_RPM))) {
       m_hopper.setVoltage(HopperConstants.LAUNCHING_VOLTAGE);
       m_column.setVoltage(ColumnConstants.LAUNCHING_VOLTAGE);
     }
@@ -162,7 +98,5 @@ public class Launch extends Command {
     m_hopper.setVoltage(0.0);
     m_column.setVoltage(0.0);
     m_hood.setPosition(HoodConstants.K_MIN_POSITION);
-
-    m_timer.stop();
   }
 }
